requirements.txt
sentinelsat
pandas
python-dotenv
requests
requests-oauthlib
oauthlib
